+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   * [Added] New method tree.visitRows()
   * [Added] New method tree.selectAll()
   * [Added] New method node.isBelowOf()
+  * [Added] `icon` option supports `{html: "..."}` content (also available for glyph-ext mapping)
   * [Changed] Re-rename clearData() to clearPersistData()
   * [Changed] #828 Re-scale debugLevel from 0:quiet to 4:verbose, allowing to suppress warnings
     and even errors.